@@ -13,7 +13,7 @@ class MyApp extends StatelessWidget {
         brightness: Brightness.light,
         primarySwatch: Colors.red,
         primaryColor: Colors.red,
-        accentColor: Colors.red,
+        colorScheme: ColorScheme.light(secondary: Colors.red),
         visualDensity: VisualDensity.adaptivePlatformDensity,
         textSelectionTheme: TextSelectionThemeData(),
       ),
@@ -21,7 +21,7 @@ class MyApp extends StatelessWidget {
         brightness: Brightness.dark,
         primarySwatch: Colors.red,
         primaryColor: Colors.red,
-        accentColor: Colors.red,
+        colorScheme: ColorScheme.dark(secondary: Colors.red),
         toggleableActiveColor: Colors.red,
         visualDensity: VisualDensity.adaptivePlatformDensity,
         textSelectionTheme: TextSelectionThemeData(),
