@@ -7,15 +7,15 @@ class S2GroupHeader extends StatelessWidget {
   final Widget title;
 
   /// Trailing widget
-  final Widget trailing;
+  final Widget? trailing;
 
   /// The header Style
-  final S2GroupHeaderStyle style;
+  final S2GroupHeaderStyle? style;
 
   /// Default constructor
   S2GroupHeader({
-    Key key,
-    @required this.title,
+    Key? key,
+    required this.title,
     this.trailing,
     this.style,
   }) : super(key: key);
@@ -23,18 +23,18 @@ class S2GroupHeader extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: style.height,
-      color: style.backgroundColor,
-      padding: style.padding,
+      height: style!.height,
+      color: style!.backgroundColor,
+      padding: style!.padding,
       child: Row(
         crossAxisAlignment:
-            style.crossAxisAlignment ?? CrossAxisAlignment.center,
+            style!.crossAxisAlignment ?? CrossAxisAlignment.center,
         mainAxisAlignment:
-            style.mainAxisAlignment ?? MainAxisAlignment.spaceBetween,
-        children: <Widget>[
+            style!.mainAxisAlignment ?? MainAxisAlignment.spaceBetween,
+        children: (<Widget?>[
           title,
           trailing,
-        ]..removeWhere((e) => e == null),
+        ]..removeWhere((e) => e == null)) as List<Widget>,
       ),
     );
   }
