@@ -7,31 +7,31 @@ import 'group_data.dart';
 
 /// A common widget builder
 typedef Widget S2WidgetBuilder<T>(
-  BuildContext context,
+  BuildContext? context,
   T value,
 );
 
 /// A common list of widget builder
 typedef List<Widget> S2ListWidgetBuilder<T>(
-  BuildContext context,
+  BuildContext? context,
   T value,
 );
 
 /// A complex builder
 typedef Widget S2ComplexWidgetBuilder<A, B>(
-  BuildContext context,
+  BuildContext? context,
   A value,
   B anotherValue,
 );
 
 /// A builder for choice list item widget
-typedef Widget S2ChoiceListBuilder<T>(S2Choice<T> choice);
+typedef Widget? S2ChoiceListBuilder<T>(S2Choice<T> choice);
 
 /// collection of builder
 @immutable
 class S2Builder<T> with Diagnosticable {
   /// Builder for custom divider widget between choices item
-  final IndexedWidgetBuilder choiceDivider;
+  final IndexedWidgetBuilder? choiceDivider;
 
   // /// Builder for progress indicator on choice load
   // final WidgetBuilder choiceProgressBuilder;
@@ -50,52 +50,52 @@ class S2Builder<T> with Diagnosticable {
 /// Builder collection of single choice widget
 class S2SingleBuilder<T> extends S2Builder<T> {
   /// Builder for custom tile widget
-  final S2WidgetBuilder<S2SingleState<T>> tile;
+  final S2WidgetBuilder<S2SingleState<T>>? tile;
 
   /// Builder for custom modal widget
-  final S2WidgetBuilder<S2SingleState<T>> modal;
+  final S2WidgetBuilder<S2SingleState<T>>? modal;
 
   /// Builder for custom modal header widget
-  final S2WidgetBuilder<S2SingleState<T>> modalHeader;
+  final S2WidgetBuilder<S2SingleState<T>>? modalHeader;
 
   /// Builder for modal filter
-  final S2WidgetBuilder<S2SingleState<T>> modalFilter;
+  final S2WidgetBuilder<S2SingleState<T>>? modalFilter;
 
   /// Builder for modal filter toggle
-  final S2WidgetBuilder<S2SingleState<T>> modalFilterToggle;
+  final S2WidgetBuilder<S2SingleState<T>>? modalFilterToggle;
 
   /// Builder for custom modal actions widget
-  final S2ListWidgetBuilder<S2SingleState<T>> modalActions;
+  final S2ListWidgetBuilder<S2SingleState<T>>? modalActions;
 
   /// Builder for modal confirm action widget
-  final S2WidgetBuilder<S2SingleState<T>> modalConfirm;
+  final S2WidgetBuilder<S2SingleState<T>>? modalConfirm;
 
   /// Builder for divider widget between header, body, and footer modal
-  final S2WidgetBuilder<S2SingleState<T>> modalDivider;
+  final S2WidgetBuilder<S2SingleState<T>>? modalDivider;
 
   /// Builder for modal footer widget
-  final S2WidgetBuilder<S2SingleState<T>> modalFooter;
+  final S2WidgetBuilder<S2SingleState<T>>? modalFooter;
 
   /// A widget builder for custom choices group
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>> group;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>>? group;
 
   /// A widget builder for custom header choices group
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>> groupHeader;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>>? groupHeader;
 
   /// Builder for each custom choices item
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choice;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choice;
 
   /// Builder for each custom choices item subtitle
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceTitle;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceTitle;
 
   /// Builder for each custom choices item subtitle
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceSubtitle;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceSubtitle;
 
   /// Builder for each custom choices item secondary
-  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceSecondary;
+  final S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceSecondary;
 
   /// Builder for custom empty display
-  final S2WidgetBuilder<S2SingleState<T>> choiceEmpty;
+  final S2WidgetBuilder<S2SingleState<T>>? choiceEmpty;
 
   /// default contructor
   const S2SingleBuilder({
@@ -115,7 +115,7 @@ class S2SingleBuilder<T> extends S2Builder<T> {
     this.choiceSubtitle,
     this.choiceSecondary,
     this.choiceEmpty,
-    IndexedWidgetBuilder choiceDivider,
+    IndexedWidgetBuilder? choiceDivider,
     // WidgetBuilder choiceProgress,
     // S2ChoiceBuilder<T> choicePager,
   }) : super(
@@ -127,23 +127,23 @@ class S2SingleBuilder<T> extends S2Builder<T> {
   /// Creates a copy of this [S2SingleBuilder] but with
   /// the given fields replaced with the new values.
   S2SingleBuilder<T> copyWith({
-    S2WidgetBuilder<S2SingleState<T>> tile,
-    S2WidgetBuilder<S2SingleState<T>> modal,
-    S2WidgetBuilder<S2SingleState<T>> modalHeader,
-    S2WidgetBuilder<S2SingleState<T>> modalFilter,
-    S2WidgetBuilder<S2SingleState<T>> modalFilterToggle,
-    S2ListWidgetBuilder<S2SingleState<T>> modalActions,
-    S2WidgetBuilder<S2SingleState<T>> modalConfirm,
-    S2WidgetBuilder<S2SingleState<T>> modalDivider,
-    S2WidgetBuilder<S2SingleState<T>> modalFooter,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>> group,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>> groupHeader,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choice,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceTitle,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceSubtitle,
-    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>> choiceSecondary,
-    S2WidgetBuilder<S2SingleState<T>> choiceEmpty,
-    IndexedWidgetBuilder choiceDivider,
+    S2WidgetBuilder<S2SingleState<T>>? tile,
+    S2WidgetBuilder<S2SingleState<T>>? modal,
+    S2WidgetBuilder<S2SingleState<T>>? modalHeader,
+    S2WidgetBuilder<S2SingleState<T>>? modalFilter,
+    S2WidgetBuilder<S2SingleState<T>>? modalFilterToggle,
+    S2ListWidgetBuilder<S2SingleState<T>>? modalActions,
+    S2WidgetBuilder<S2SingleState<T>>? modalConfirm,
+    S2WidgetBuilder<S2SingleState<T>>? modalDivider,
+    S2WidgetBuilder<S2SingleState<T>>? modalFooter,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>>? group,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Group<T>>? groupHeader,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choice,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceTitle,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceSubtitle,
+    S2ComplexWidgetBuilder<S2SingleState<T>, S2Choice<T>>? choiceSecondary,
+    S2WidgetBuilder<S2SingleState<T>>? choiceEmpty,
+    IndexedWidgetBuilder? choiceDivider,
     // WidgetBuilder choiceProgress,
     // S2ChoiceBuilder<T> choicePager,
   }) {
@@ -172,7 +172,7 @@ class S2SingleBuilder<T> extends S2Builder<T> {
 
   /// Creates a copy of this [S2SingleBuilder] but with
   /// the given fields replaced with the new values.
-  S2SingleBuilder<T> merge(S2SingleBuilder<T> other) {
+  S2SingleBuilder<T> merge(S2SingleBuilder<T>? other) {
     // if null return current object
     if (other == null) return this;
 
@@ -203,52 +203,52 @@ class S2SingleBuilder<T> extends S2Builder<T> {
 /// Builder collection of multiple choice widget
 class S2MultiBuilder<T> extends S2Builder<T> {
   /// Builder for custom tile widget
-  final S2WidgetBuilder<S2MultiState<T>> tile;
+  final S2WidgetBuilder<S2MultiState<T>>? tile;
 
   /// Builder for custom modal widget
-  final S2WidgetBuilder<S2MultiState<T>> modal;
+  final S2WidgetBuilder<S2MultiState<T>>? modal;
 
   /// Builder for custom modal header widget
-  final S2WidgetBuilder<S2MultiState<T>> modalHeader;
+  final S2WidgetBuilder<S2MultiState<T>>? modalHeader;
 
   /// Builder for modal filter
-  final S2WidgetBuilder<S2MultiState<T>> modalFilter;
+  final S2WidgetBuilder<S2MultiState<T>>? modalFilter;
 
   /// Builder for modal filter toggle
-  final S2WidgetBuilder<S2MultiState<T>> modalFilterToggle;
+  final S2WidgetBuilder<S2MultiState<T>>? modalFilterToggle;
 
   /// Builder for custom modal actions widget
-  final S2ListWidgetBuilder<S2MultiState<T>> modalActions;
+  final S2ListWidgetBuilder<S2MultiState<T>>? modalActions;
 
   /// Builder for custom modal confirm action widget
-  final S2WidgetBuilder<S2MultiState<T>> modalConfirm;
+  final S2WidgetBuilder<S2MultiState<T>>? modalConfirm;
 
   /// Builder for divider widget between header, body, and footer modal
-  final S2WidgetBuilder<S2MultiState<T>> modalDivider;
+  final S2WidgetBuilder<S2MultiState<T>>? modalDivider;
 
   /// Builder for custom modal footer widget
-  final S2WidgetBuilder<S2MultiState<T>> modalFooter;
+  final S2WidgetBuilder<S2MultiState<T>>? modalFooter;
 
   /// A widget builder for custom choices group
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>> group;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>>? group;
 
   /// A widget builder for custom header choices group
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>> groupHeader;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>>? groupHeader;
 
   /// Builder for each custom choices item
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choice;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choice;
 
   /// Builder for each custom choices item subtitle
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceTitle;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceTitle;
 
   /// Builder for each custom choices item subtitle
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceSubtitle;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceSubtitle;
 
   /// Builder for each custom choices item secondary
-  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceSecondary;
+  final S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceSecondary;
 
   /// Builder for custom empty display
-  final S2WidgetBuilder<S2MultiState<T>> choiceEmpty;
+  final S2WidgetBuilder<S2MultiState<T>>? choiceEmpty;
 
   /// default constructor
   const S2MultiBuilder({
@@ -268,7 +268,7 @@ class S2MultiBuilder<T> extends S2Builder<T> {
     this.choiceSubtitle,
     this.choiceSecondary,
     this.choiceEmpty,
-    IndexedWidgetBuilder choiceDivider,
+    IndexedWidgetBuilder? choiceDivider,
     // WidgetBuilder choiceProgress,
     // S2ChoiceBuilder<T> choicePager,
   }) : super(
@@ -280,23 +280,23 @@ class S2MultiBuilder<T> extends S2Builder<T> {
   /// Creates a copy of this [S2MultiBuilder] but with
   /// the given fields replaced with the new values.
   S2MultiBuilder<T> copyWith({
-    S2WidgetBuilder<S2MultiState<T>> tile,
-    S2WidgetBuilder<S2MultiState<T>> modal,
-    S2WidgetBuilder<S2MultiState<T>> modalHeader,
-    S2WidgetBuilder<S2MultiState<T>> modalFilter,
-    S2WidgetBuilder<S2MultiState<T>> modalFilterToggle,
-    S2ListWidgetBuilder<S2MultiState<T>> modalActions,
-    S2WidgetBuilder<S2MultiState<T>> modalConfirm,
-    S2WidgetBuilder<S2MultiState<T>> modalDivider,
-    S2WidgetBuilder<S2MultiState<T>> modalFooter,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>> group,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>> groupHeader,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choice,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceTitle,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceSubtitle,
-    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>> choiceSecondary,
-    S2WidgetBuilder<S2MultiState<T>> choiceEmpty,
-    IndexedWidgetBuilder choiceDivider,
+    S2WidgetBuilder<S2MultiState<T>>? tile,
+    S2WidgetBuilder<S2MultiState<T>>? modal,
+    S2WidgetBuilder<S2MultiState<T>>? modalHeader,
+    S2WidgetBuilder<S2MultiState<T>>? modalFilter,
+    S2WidgetBuilder<S2MultiState<T>>? modalFilterToggle,
+    S2ListWidgetBuilder<S2MultiState<T>>? modalActions,
+    S2WidgetBuilder<S2MultiState<T>>? modalConfirm,
+    S2WidgetBuilder<S2MultiState<T>>? modalDivider,
+    S2WidgetBuilder<S2MultiState<T>>? modalFooter,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>>? group,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Group<T>>? groupHeader,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choice,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceTitle,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceSubtitle,
+    S2ComplexWidgetBuilder<S2MultiState<T>, S2Choice<T>>? choiceSecondary,
+    S2WidgetBuilder<S2MultiState<T>>? choiceEmpty,
+    IndexedWidgetBuilder? choiceDivider,
     // WidgetBuilder choiceProgress,
     // S2ChoiceBuilder<T> choicePager,
   }) {
@@ -325,7 +325,7 @@ class S2MultiBuilder<T> extends S2Builder<T> {
 
   /// Creates a copy of this [S2MultiBuilder] but with
   /// the given fields replaced with the new values.
-  S2MultiBuilder<T> merge(S2MultiBuilder<T> other) {
+  S2MultiBuilder<T> merge(S2MultiBuilder<T>? other) {
     // if null return current object
     if (other == null) return this;
 
