@@ -15,10 +15,10 @@ class S2Choice<T> with Diagnosticable {
   final String title;
 
   /// Represent as secondary text
-  final String subtitle;
+  final String? subtitle;
 
   /// The choice will grouped by this property value
-  final String group;
+  final String? group;
 
   /// Whether the choice is disabled or enabled
   final bool disabled;
@@ -30,21 +30,21 @@ class S2Choice<T> with Diagnosticable {
   final dynamic meta;
 
   /// Individual unselected choice item style
-  final S2ChoiceStyle style;
+  final S2ChoiceStyle? style;
 
   /// Individual selected choice item style
-  final S2ChoiceStyle activeStyle;
+  final S2ChoiceStyle? activeStyle;
 
   /// Callback to select choice
-  final ValueSetter<bool> select;
+  final ValueSetter<bool?>? select;
 
   /// Whether the choice is selected or not
   final bool selected;
 
   /// Default constructor
   S2Choice({
-    @required this.value,
-    @required this.title,
+    required this.value,
+    required this.title,
     this.subtitle,
     this.group,
     this.disabled = false,
@@ -54,29 +54,28 @@ class S2Choice<T> with Diagnosticable {
     this.activeStyle,
     this.select,
     this.selected = false,
-  })  : assert(disabled != null),
-        assert(hidden != null);
+  });
 
   /// Helper to create option list from any list
   static List<S2Choice<R>> listFrom<R, E>({
-    @required List<E> source,
-    @required _S2OptionProp<E, R> value,
-    @required _S2OptionProp<E, String> title,
-    _S2OptionProp<E, String> subtitle,
-    _S2OptionProp<E, String> group,
-    _S2OptionProp<E, bool> disabled,
-    _S2OptionProp<E, bool> hidden,
-    _S2OptionProp<E, dynamic> meta,
-    _S2OptionProp<E, S2ChoiceStyle> style,
-    _S2OptionProp<E, S2ChoiceStyle> activeStyle,
+    required List<E> source,
+    required _S2OptionProp<E, R> value,
+    required _S2OptionProp<E, String> title,
+    _S2OptionProp<E, String>? subtitle,
+    _S2OptionProp<E, String>? group,
+    _S2OptionProp<E, bool>? disabled,
+    _S2OptionProp<E, bool>? hidden,
+    _S2OptionProp<E, dynamic>? meta,
+    _S2OptionProp<E, S2ChoiceStyle>? style,
+    _S2OptionProp<E, S2ChoiceStyle>? activeStyle,
   }) =>
       source
           .asMap()
           .map((index, item) => MapEntry(
               index,
               S2Choice<R>(
-                value: value?.call(index, item),
-                title: title?.call(index, item),
+                value: value.call(index, item),
+                title: title.call(index, item),
                 subtitle: subtitle?.call(index, item),
                 group: group?.call(index, item),
                 disabled: disabled?.call(index, item) ?? false,
@@ -94,7 +93,7 @@ class S2Choice<T> with Diagnosticable {
         _testPropBy(group, query);
   }
 
-  bool _testPropBy(String prop, String query) {
+  bool _testPropBy(String? prop, String query) {
     return prop != null ? normalized(prop).contains(normalized(query)) : false;
   }
 
@@ -108,22 +107,22 @@ class S2Choice<T> with Diagnosticable {
   @override
   int get hashCode => value.hashCode;
 
-  S2ChoiceStyle get effectiveStyle => selected == true ? activeStyle : style;
+  S2ChoiceStyle? get effectiveStyle => selected == true ? activeStyle : style;
 
   /// Creates a copy of this [S2Choice] but with
   /// the given fields replaced with the new values.
   S2Choice<T> copyWith({
-    T value,
-    String title,
-    String subtitle,
-    String group,
-    bool disabled,
-    bool hidden,
+    T? value,
+    String? title,
+    String? subtitle,
+    String? group,
+    bool? disabled,
+    bool? hidden,
     dynamic meta,
-    S2ChoiceStyle style,
-    S2ChoiceStyle activeStyle,
-    ValueSetter<bool> select,
-    bool selected,
+    S2ChoiceStyle? style,
+    S2ChoiceStyle? activeStyle,
+    ValueSetter<bool?>? select,
+    bool? selected,
   }) {
     return S2Choice<T>(
       value: value ?? this.value,
@@ -143,9 +142,6 @@ class S2Choice<T> with Diagnosticable {
   /// Creates a copy of this [S2Choice] but with
   /// the given fields replaced with the new values.
   S2Choice<T> merge(S2Choice<T> other) {
-    // if null return current object
-    if (other == null) return this;
-
     return copyWith(
       value: other.value,
       title: other.title,
