@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import '../model/choice_item.dart';
 import '../model/group_data.dart';
 import '../model/group_config.dart';
@@ -16,19 +15,17 @@ class S2Choices<T> {
 
   /// default constructor
   S2Choices({
-    @required this.items,
-    @required this.query,
-    @required this.groupConfig,
+    required this.items,
+    required this.query,
+    required this.groupConfig,
   });
 
   /// return a filtered list of options
   List<S2Choice<T>> get filteredItems {
-    return query != null
-        ? nonHiddenItems
-            .where((S2Choice<T> item) => item.contains(query))
-            .toList()
-            .cast<S2Choice<T>>()
-        : nonHiddenItems;
+    return nonHiddenItems
+        .where((S2Choice<T> item) => item.contains(query))
+        .toList()
+        .cast<S2Choice<T>>();
   }
 
   /// return a non hidden option item
@@ -52,7 +49,7 @@ class S2Choices<T> {
 
     // sort the list when the comparator is provided
     if (groupConfig.sortBy != null)
-      return groups..sort(groupConfig.sortBy.compare);
+      return groups..sort(groupConfig.sortBy!.compare);
 
     return groups;
   }
@@ -75,6 +72,6 @@ class S2Choices<T> {
 
   /// whether the list need to be grouped or not
   bool get isGrouped {
-    return groupConfig.enabled && groupKeys != null && groupKeys.isNotEmpty;
+    return groupConfig.enabled && groupKeys.isNotEmpty;
   }
 }
