@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
 import './group_style.dart';
 import './group_sort.dart';
 
@@ -18,7 +17,7 @@ class S2GroupConfig with Diagnosticable {
   /// If [enabled] is `true`, comparator function to sort the group keys
   ///
   /// Defaults to `null`, and it means disabled the sorting
-  final S2GroupSort sortBy;
+  final S2GroupSort? sortBy;
 
   /// Configure choices group header theme
   final S2GroupHeaderStyle headerStyle;
@@ -30,8 +29,7 @@ class S2GroupConfig with Diagnosticable {
     this.useCounter = true,
     this.sortBy,
     this.headerStyle = const S2GroupHeaderStyle(),
-  })  : assert(enabled != null),
-        assert(headerStyle != null);
+  });
 
   /// Opposite value of enabled
   bool get disabled => !enabled;
@@ -39,24 +37,24 @@ class S2GroupConfig with Diagnosticable {
   /// Creates a copy of this [S2GroupConfig] but with
   /// the given fields replaced with the new values.
   S2GroupConfig copyWith({
-    bool enabled,
-    bool useSelector,
-    bool useCounter,
-    S2GroupSort sortBy,
-    S2GroupHeaderStyle headerStyle,
+    bool? enabled,
+    bool? useSelector,
+    bool? useCounter,
+    S2GroupSort? sortBy,
+    S2GroupHeaderStyle? headerStyle,
   }) {
     return S2GroupConfig(
       enabled: enabled ?? this.enabled,
       useSelector: useSelector ?? this.useSelector,
       useCounter: useCounter ?? this.useCounter,
       sortBy: sortBy ?? this.sortBy,
-      headerStyle: this.headerStyle?.merge(headerStyle) ?? headerStyle,
+      headerStyle: this.headerStyle.merge(headerStyle),
     );
   }
 
   /// Returns a new [S2GroupConfig] that is
   /// a combination of this object and the given [other] style.
-  S2GroupConfig merge(S2GroupConfig other) {
+  S2GroupConfig merge(S2GroupConfig? other) {
     // if null return current object
     if (other == null) return this;
 
