@@ -5,12 +5,12 @@ import '../model/choice_item.dart';
 typedef Future<S2Choice<T>> S2SingleSelectedResolver<T>(T value);
 
 /// Function to return a `List` of [S2Choice] from a `List` of `value`
-typedef Future<List<S2Choice<T>>> S2MultiSelectedResolver<T>(List<T> value);
+typedef Future<List<S2Choice<T>>> S2MultiSelectedResolver<T>(List<T>? value);
 
 /// State of the selected choice
 abstract class S2Selected<T> extends ChangeNotifier {
   /// A `String` to return in `toString` if the `title` is empty
-  String placeholder;
+  String? placeholder;
 
   /// Function to resolve [choice] from [value]
   covariant var resolver;
@@ -53,7 +53,7 @@ abstract class S2Selected<T> extends ChangeNotifier {
   String error = '';
 
   /// Returns `true` when the selected is valid
-  bool get isValid => error == null || error?.length == 0;
+  bool get isValid => error.length == 0;
 
   /// Returns `true` when the selected is not valid
   bool get isNotValid => isValid != true;
@@ -93,29 +93,29 @@ abstract class S2Selected<T> extends ChangeNotifier {
 class S2SingleSelected<T> extends S2Selected<T> {
   /// Default constructor
   S2SingleSelected({
-    T value,
-    S2Choice<T> choice,
+    T? value,
+    S2Choice<T>? choice,
     this.resolver,
     this.validation,
     this.placeholder,
   })  : _value = value,
         _choice = choice;
 
-  T _value;
+  T? _value;
 
-  S2Choice<T> _choice;
+  S2Choice<T?>? _choice;
 
   /// Function to resolve [choice] from [value]
   @override
-  S2SingleSelectedResolver<T> resolver;
+  S2SingleSelectedResolver<T?>? resolver;
 
   /// A function used to validate the selection
   @override
-  final S2Validation<S2Choice<T>> validation;
+  final S2Validation<S2Choice<T>>? validation;
 
   /// a `String` to return in `toString` if the `title` is empty
   @override
-  final String placeholder;
+  final String? placeholder;
 
   @override
   int get length => choice != null ? 1 : 0;
@@ -128,7 +128,7 @@ class S2SingleSelected<T> extends S2Selected<T> {
 
   @override
   void resolve({
-    S2SingleSelectedResolver<T> defaultResolver,
+    S2SingleSelectedResolver<T?>? defaultResolver,
   }) async {
     if (isResolved) return null;
 
@@ -147,14 +147,14 @@ class S2SingleSelected<T> extends S2Selected<T> {
   }
 
   @override
-  set choice(S2Choice<T> val) {
+  set choice(S2Choice<T?>? val) {
     _choice = val;
     _value = null;
     validate();
   }
 
   @override
-  set value(T val) {
+  set value(T? val) {
     _value = val;
     _choice = null;
     resolve();
@@ -162,31 +162,31 @@ class S2SingleSelected<T> extends S2Selected<T> {
 
   /// return a single selected [S2Choice]
   @override
-  S2Choice<T> get choice {
+  S2Choice<T?>? get choice {
     return _choice;
   }
 
   /// return [choice.value]
   @override
-  T get value {
+  T? get value {
     return choice?.value ?? _value;
   }
 
   /// return [choice.title]
   @override
-  String get title {
+  String? get title {
     return choice?.title;
   }
 
   /// return [choice.subtitle]
   @override
-  String get subtitle {
+  String? get subtitle {
     return choice?.subtitle;
   }
 
   /// return [choice.group]
   @override
-  String get group {
+  String? get group {
     return choice?.group;
   }
 
@@ -204,28 +204,28 @@ class S2SingleSelected<T> extends S2Selected<T> {
 class S2MultiSelected<T> extends S2Selected<T> {
   /// Default Constructor
   S2MultiSelected({
-    List<T> value,
-    List<S2Choice<T>> choice,
+    List<T>? value,
+    List<S2Choice<T>>? choice,
     this.resolver,
     this.validation,
     this.placeholder,
   })  : _value = List<T>.from(value ?? []),
         _choice = choice != null ? List<S2Choice<T>>.from(choice) : null;
 
-  List<T> _value;
+  List<T>? _value;
 
-  List<S2Choice<T>> _choice;
+  List<S2Choice<T>>? _choice;
 
   /// Function to resolve [choice] from [value]
   @override
-  S2MultiSelectedResolver<T> resolver;
+  S2MultiSelectedResolver<T>? resolver;
 
   /// A function used to validate the selection
-  final S2Validation<List<S2Choice<T>>> validation;
+  final S2Validation<List<S2Choice<T>>>? validation;
 
   /// a `String` to return in `toString` if the `title` is empty
   @override
-  final String placeholder;
+  final String? placeholder;
 
   @override
   int get length => choice?.length ?? 0;
@@ -238,7 +238,7 @@ class S2MultiSelected<T> extends S2Selected<T> {
 
   @override
   void resolve({
-    S2MultiSelectedResolver<T> defaultResolver,
+    S2MultiSelectedResolver<T>? defaultResolver,
   }) async {
     if (isResolved) return null;
 
@@ -257,14 +257,14 @@ class S2MultiSelected<T> extends S2Selected<T> {
   }
 
   @override
-  set choice(List<S2Choice<T>> choices) {
+  set choice(List<S2Choice<T>>? choices) {
     _choice = List<S2Choice<T>>.from(choices ?? []);
     _value = null;
     validate();
   }
 
   @override
-  set value(List<T> val) {
+  set value(List<T>? val) {
     _value = List<T>.from(val ?? []);
     _choice = null;
     resolve();
@@ -272,39 +272,39 @@ class S2MultiSelected<T> extends S2Selected<T> {
 
   /// return an array of the selected [S2Choice]
   @override
-  List<S2Choice<T>> get choice {
+  List<S2Choice<T>>? get choice {
     return _choice;
   }
 
   /// return an array of `value` of the selected [choice]
   @override
-  List<T> get value {
+  List<T>? get value {
     return isNotEmpty
-        ? choice.map((S2Choice<T> item) => item.value).toList()
+        ? choice!.map((S2Choice<T> item) => item.value).toList()
         : _value;
   }
 
   /// return an array of `title` of the selected [choice]
   @override
-  List<String> get title {
+  List<String>? get title {
     return isNotEmpty
-        ? choice.map((S2Choice<T> item) => item.title).toList()
+        ? choice!.map((S2Choice<T> item) => item.title).toList()
         : null;
   }
 
   /// return an array of `subtitle` of the selected [choice]
   @override
-  List<String> get subtitle {
+  List<String?>? get subtitle {
     return isNotEmpty
-        ? choice.map((S2Choice<T> item) => item.subtitle).toList()
+        ? choice!.map((S2Choice<T> item) => item.subtitle).toList()
         : null;
   }
 
   /// return an array of `group` of the selected [choice]
   @override
-  List<String> get group {
+  List<String?>? get group {
     return isNotEmpty
-        ? choice.map((S2Choice<T> item) => item.group).toList()
+        ? choice!.map((S2Choice<T> item) => item.group).toList()
         : null;
   }
 
