@@ -3,23 +3,23 @@ import 'package:flutter/widgets.dart';
 /// Text widget that supports highlight
 class S2Text extends StatelessWidget {
   /// The text data string
-  final String text;
+  final String? text;
 
   /// The text style
-  final TextStyle style;
+  final TextStyle? style;
 
   /// The highlight data string
-  final String highlight;
+  final String? highlight;
 
   /// The highlight color
-  final Color highlightColor;
+  final Color? highlightColor;
 
   /// Whether the match is case sensitive or not
   final bool caseSensitive;
 
   /// Default constructor
   const S2Text({
-    Key key,
+    Key? key,
     this.text,
     this.style,
     this.highlight,
@@ -34,33 +34,33 @@ class S2Text extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if ((highlight?.isEmpty ?? true) || text.isEmpty) {
-      return Text(text, style: style);
+    if ((highlight?.isEmpty ?? true) || text!.isEmpty) {
+      return Text(text!, style: style);
     }
 
     final TextStyle defaultTextStyle = DefaultTextStyle.of(context).style;
     final TextStyle textStyle = defaultTextStyle.merge(style);
-    final Pattern pattern = RegExp(highlight, caseSensitive: caseSensitive);
+    final Pattern pattern = RegExp(highlight!, caseSensitive: caseSensitive);
     int start = 0;
     int indexOfHighlight;
     List<TextSpan> spans = [];
 
     do {
-      indexOfHighlight = text.indexOf(pattern, start);
+      indexOfHighlight = text!.indexOf(pattern, start);
       if (indexOfHighlight < 0) {
         // no highlight
-        final content = text.substring(start, text.length);
+        final content = text!.substring(start, text!.length);
         spans.add(_normalSpan(content, textStyle));
         break;
       }
       if (indexOfHighlight == start) {
         // start with highlight.
-        final highlightedText = text.substring(start, start + highlight.length);
+        final highlightedText = text!.substring(start, start + highlight!.length);
         spans.add(_highlightSpan(highlightedText, textStyle));
         start += highlightedText.length;
       } else {
         // normal + highlight
-        final String normalText = text.substring(start, indexOfHighlight);
+        final String normalText = text!.substring(start, indexOfHighlight);
         spans.add(_normalSpan(normalText, textStyle));
         start += normalText.length;
       }
