@@ -10,16 +10,16 @@ class S2TileChips extends StatelessWidget {
   final IndexedWidgetBuilder chipLabelBuilder;
 
   /// Widget builder for chip avatar item
-  final IndexedWidgetBuilder chipAvatarBuilder;
+  final IndexedWidgetBuilder? chipAvatarBuilder;
 
   /// Widget builder for chip item
-  final IndexedWidgetBuilder chipBuilder;
+  final IndexedWidgetBuilder? chipBuilder;
 
   /// Called when the user delete the chip item.
-  final ValueChanged<int> chipOnDelete;
+  final ValueChanged<int>? chipOnDelete;
 
   /// The primary color of the chip item
-  final Color chipColor;
+  final Color? chipColor;
 
   /// Whether the chip is outlined or not
   final bool chipOutlined;
@@ -29,40 +29,40 @@ class S2TileChips extends StatelessWidget {
 
   /// If [chipOutlined] is [true] this value becomes the border opacity,
   /// If [chipOutlined] is [false] this value becomes the background opacity
-  final double chipOpacity;
+  final double? chipOpacity;
 
   /// The icon of the delete button
-  final Icon chipDeleteIcon;
+  final Icon? chipDeleteIcon;
 
   /// Chip spacing
-  final double chipSpacing;
+  final double? chipSpacing;
 
   /// Chip run spacing
-  final double chipRunSpacing;
+  final double? chipRunSpacing;
 
   /// Chip shape border
-  final ShapeBorder chipShape;
+  final ShapeBorder? chipShape;
 
   /// The [TextStyle] of the chip label
-  final TextStyle chipLabelStyle;
+  final TextStyle? chipLabelStyle;
 
   /// The elevation of the chip widget
   final double chipElevation;
 
   /// The [Widget] displayed when the [values] is null
-  final Widget placeholder;
+  final Widget? placeholder;
 
   /// Whether the chip list is scrollable or not
   final bool scrollable;
 
   /// Chip list padding
-  final EdgeInsetsGeometry padding;
+  final EdgeInsetsGeometry? padding;
 
   /// Create a chips tile/trigger widget
   S2TileChips({
-    Key key,
-    @required this.chipLength,
-    @required this.chipLabelBuilder,
+    Key? key,
+    required this.chipLength,
+    required this.chipLabelBuilder,
     this.chipAvatarBuilder,
     this.chipBuilder,
     this.chipOnDelete,
@@ -79,10 +79,7 @@ class S2TileChips extends StatelessWidget {
     this.placeholder,
     this.scrollable = false,
     this.padding,
-  })  : assert(chipOutlined != null),
-        assert(chipRaised != null),
-        assert(scrollable != null),
-        super(key: key);
+  }) : super(key: key);
 
   /// default padding
   static const EdgeInsetsGeometry defaultPadding =
@@ -164,10 +161,10 @@ class S2TileChips extends StatelessWidget {
 
   Widget _chipGenerator(BuildContext context, int i) {
     return Chip(
-      label: chipLabelBuilder?.call(context, i),
+      label: chipLabelBuilder.call(context, i),
       avatar: chipAvatarBuilder?.call(context, i),
       deleteIcon: chipDeleteIcon,
-      onDeleted: chipOnDelete != null ? () => chipOnDelete(i) : null,
+      onDeleted: chipOnDelete != null ? () => chipOnDelete!(i) : null,
     );
   }
 }
