@@ -5,7 +5,8 @@ import 'choices.dart' as choices;
 
 void main() {
   testSmartSelect(
-    title: 'Full page modal, default tile and radio choices displayed correctly',
+    title:
+        'Full page modal, default tile and radio choices displayed correctly',
     initialChoice: null,
     choiceToSelect: choices.days[3],
     choiceItems: choices.days,
@@ -14,7 +15,8 @@ void main() {
   );
 
   testSmartSelect(
-    title: 'Bottomsheet modal, default tile and chips choices displayed correctly',
+    title:
+        'Bottomsheet modal, default tile and chips choices displayed correctly',
     placeholder: 'Pilih Salah Satu',
     initialChoice: null,
     choiceToSelect: choices.heroes[2],
@@ -24,7 +26,8 @@ void main() {
   );
 
   testSmartSelect(
-    title: 'Popup dialog modal, default tile and switch choices displayed correctly',
+    title:
+        'Popup dialog modal, default tile and switch choices displayed correctly',
     initialChoice: choices.frameworks[0],
     choiceToSelect: choices.frameworks[1],
     choiceItems: choices.frameworks,
@@ -34,15 +37,15 @@ void main() {
 }
 
 testSmartSelect<T>({
-  @required String title,
-  @required S2Choice<T> initialChoice,
-  @required S2Choice<T> choiceToSelect,
-  @required List<S2Choice<T>> choiceItems,
+  required String title,
+  required S2Choice<T>? initialChoice,
+  required S2Choice<T> choiceToSelect,
+  required List<S2Choice<T>> choiceItems,
   S2ModalType modalType = S2ModalType.fullPage,
-  S2ChoiceType choiceType,
+  S2ChoiceType? choiceType,
   String placeholder = 'Select one',
 }) {
-  S2Choice<T> selectedChoice = initialChoice;
+  S2Choice<T?>? selectedChoice = initialChoice;
 
   testWidgets(title, (WidgetTester tester) async {
     await tester.pumpWidget(
@@ -54,7 +57,7 @@ testSmartSelect<T>({
           choiceItems: choiceItems,
           modalType: modalType,
           choiceType: choiceType,
-          onChange: (selected) => selectedChoice = selected.choice,
+          onChange: (selected) => selectedChoice = selected?.choice,
         ),
       ),
     );
@@ -69,21 +72,21 @@ testSmartSelect<T>({
       reason: 'SmartSelect widget is displayed',
     );
 
-    final s2State = tester.state<S2SingleState<T>>(s2Finder);
+    final s2State = tester.state<S2SingleState<T?>>(s2Finder);
     expect(
-      s2State.selected.choice,
+      s2State.selected!.choice,
       initialChoice,
       reason: 'Initial choice is correct',
     );
     expect(
-      s2State.selected.value,
+      s2State.selected!.value,
       initialChoice?.value,
       reason: 'Initial value is correct',
     );
 
     final tileFinder = find.descendant(
       of: s2Finder,
-      matching: find.byWidgetPredicate((widget) => widget is S2Tile<T>),
+      matching: find.byWidgetPredicate((widget) => widget is S2Tile<T?>),
     );
     expect(tileFinder, findsOneWidget, reason: 'Trigger tile displayed');
 
@@ -167,7 +170,7 @@ testSmartSelect<T>({
     await tester.pumpAndSettle();
 
     expect(
-      s2State.selected.choice,
+      s2State.selected!.choice,
       choiceToSelect,
       reason: 'New selected choice to internal choice is correct',
     );
@@ -178,12 +181,12 @@ testSmartSelect<T>({
     );
 
     expect(
-      s2State.selected.value,
+      s2State.selected!.value,
       choiceToSelect.value,
       reason: 'New selected value to internal value is correct',
     );
     expect(
-      selectedChoice.value,
+      selectedChoice!.value,
       choiceToSelect.value,
       reason: 'New selected value to external value is correct',
     );
@@ -194,8 +197,8 @@ class Bootstrap extends StatelessWidget {
   final Widget child;
 
   const Bootstrap({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
   }) : super(key: key);
 
   @override
